Work Session #5
In the final Work Session, I will add comments on the code, but for now, I will explain here. I decided that it would be much too hard for the user to select a certain team, so I decided that instead, the user would select any name of their choosing and the draft would operate with a guaranteed method of how it would run (minimizing the if statements needed). However, since there are 10 rounds, there is still a crazy amount of code that was mostly copying and pasting each round. I tested the draft and it works. A full explanation of the game and the draft specifically will be on the diary.