Work Session #6
I heavily simplified the code here, making a function for every single round with three parameters that lessened how much code there was. I also added the functions to generate the score of a game against two teams. However, it is not complete yet, and I will submit my full final code with comments included in the Haiku dropbox.